requirements.txt
google-generativeai
flask
flask_cors
pillow
pypdf2
